label saving refactor & testing
